+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.3.3 - 2026-01-31
+
+### Fixed
+- Running tests with an active filter now only runs the filtered tests (not all tests)
+- If tests are selected with a filter active, only the selected tests run
+
 ## v0.3.2 - 2026-01-31
 
 ### Added